Add input event for BTHome button presses
All of the BTHome devices seem to have a button for setup and
configuration.  For some devices, like the Shelly BLU Button1, this
button is really the only function of the device.  For others, like the
BLU H&T, the button is still useful for driving automations.

This change simply defines an input and the corresponding button events
for all of the BTHome devices that can be registered with the system.