Add BLU button event and Shelly BLU Button1 support (#668)
* Add input event for BTHome button presses

All of the BTHome devices seem to have a button for setup and
configuration.  For some devices, like the Shelly BLU Button1, this
button is really the only function of the device.  For others, like the
BLU H&T, the button is still useful for driving automations.

This change simply defines an input and the corresponding button events
for all of the BTHome devices that can be registered with the system.

* Adds support for the Shelly BLU Button1

This change simply defines the constants necessary to recognize the BLU
Button1 as a valid and supported BTHome device.

* Address PR feedback

* Update python_scripts/shellies_discovery_gen2.py

* Update python_scripts/shellies_discovery_gen2.py

* Update python_scripts/shellies_discovery_gen2.py

* Update device id prefix, per review

* Update python_scripts/shellies_discovery_gen2.py

* Fix logic

* Cleaning

---------

Co-authored-by: Maciej Bieniek <bieniu@users.noreply.github.com>